// src/app/shared/models/api/api-movie-details.ts
export interface ApiMovieDetails {
    adult: boolean;
    backdrop_path: string | null;
    budget: number;
    homepage: string;
    id: number;
    imdb_id: string;
    original_language: string;
    original_title: string;
    overview: string;
    popularity: number;
    poster_path: string;
    release_date: string;
    revenue: number;
    status: string;
    tagline: string;
    title: string;
    vote_average: number;
    vote_count: number;
}